[Scribe] Rework the Bot class a bit
The use of BackgroundWebSocket was originally a kludge to be able to
use the socket's recv() method as a timer (which would break if one
used it directly whilst SSL was on [1]). After removing EventScheduler
support from Bot itself (as has not been completed in the previous
commit) because the class' functionality does not depend on it (and
running a minimal bot does not require two threads anymore), it
becomes unnecessary.
The (BackgroundWebSocket) class will be removed once Scribe itself is
rewritten.

[1] See commit f123f4f9d7337174d622d2a261170f3a7d429434