[Frontend] Improve popup icon preloading
Previously, if the page stayed open for long, the display of a popup
would trigger a server request for the images. To avoid that, the
icons are preloaded into <OBJECT> elements, whose content is processed
into data: URI-s, which are swapped in instead of the original URL-s.
The unusual approach is chosen to avoid using XMLHttpRequest-s, which
may be logged by the browser developer tools. :P