[Frontend] Make the user list animation work in the background
Recent versions of Firefox started delaying animations commenced while
a tab is not visible to the moment when it is shown again. While this
may be useful in some cases (and is actively abused by the UI message
flashing), it is not for the user list, where the animation is meant to
measure when a user had been last active.