[Backend] Implement client-defined connection timeouts
Clients with potentially patchy connectivity (for example, mobile ones)
can send a "next" time along with a ping, which represents a promise to
send another ping not after the given time (the backend applies a grace
time of one second to account for connection latency); the backend will
terminate a connection if the promised ping should not arrive. Messages
other than pings do not postpone the deadline. Clients should, in order
to detect dropped connections reliably, also check whether responses to
their pings arrive. The feature remains opt-in to allow clients such as
Scribe to omit its use if not necessary.

Also, the text block above is weirdly neatly justified.