[Backend] Fix UniqueCounter sequence resetting
Due to an incorrect idiom, the former implementation would reset the
sequence counter whenever there was *a second of difference to the
last invocation* instead of *when the integral second changed*.
Although both approaches should be effective against the leap second
problem (i.e. potential duplicates during those -- monotonicity is
still, alas, not given), the latter one is the intended behavior.