[Backend] Add console method for synchronous I/O
The new method allows users of the console to identify the result of a
command execution in the output stream; this can be used to consume the
output stream exactly up to the last command's result (while also
capturing things going to the output stream directly). Attempting to do
that while using runCommand() may result in race conditions.