[Backend] Rework Lexer state and pattern handling
There is now a single StandardState class used at compile and run time.
In order not to keep redundant pattern lists and sets, patterns are
indexed by their name (which now must be unique). This allows
simplifying the matcher management; in particular, now, matchers are
kept around for all patterns (with lazy creation), assuming that they
will be reused.