[Frontend] Avoid round-off errors in scrollIntoViewEx
Many browsers coerce the scrollTop property to integer values upon
assignment, resulting in the input bar "jiggling around" while being
moved upwards by pressing and holding the Up arrow key. This happened
because the scrollTop property was floor-rounded, causing the input
bar to be off by one pixel in some situations. The new code detects
this behavior, corrects it, and sets a flag to always preventively
round the value to be scrolled to. Detecting that the flag needs to
be cleared is not implemented, and might become a follow-up issue.