[Backend] Fix WeakIndex bug
Removing V-s from a Collection of WeakReference<V>-s is not going to
help. Oh, and all the V-s are null by the time the removal is
attempted.

Also changes the type of toRemove to a set for better removeAll()
performance (the AbstractCollection implementation, necessarily,
iterates over its receiver and invokes the parameter's contains()
for each element).