[Frontend] Avoid settings button border width mismatch
At zoom levels where CSS pixels correspond to fractional device pixels
(including high-DPI displays), integer rounding could cause
misalignment of the inner edges of the settings button's and popup's
borders. This is fixed by increasing the (nonzero parts of the) border
width of .settings.visible. As a side effect, the outline of the
settings button now extends outside the settings popup when the popup
is open.

A follow-up issue is that, because of the same integer rounding, the
settings icon may slightly change its position between the popup-visible
and popup-hidden states. A more complex fix may be required.