[Frontend] Avoid user list menu pixel rounding issues
As border widths are pixel-rounded while paddings are not, "border
pixels" do not equal "padding pixels" in general, and therefore should
not be equated. In particular, user list nicks gain a transparent border
instead of part of their padding, with the border-color changing when
a nick is in "menu mode". In oder to avoid the background "cutting into"
the menu's border at the corners, only the relevant parts of the
nickname padding are replaced with a border, which may -- alas -- result
in subtly unbalanced visible paddings.

This is Part 4 of the HiDPI bug fix commit series.