[Frontend] Tweak settings button placeholder style
As border pixels do not equal padding pixels, the placeholder must have
the right amount of either kind, or the rest of the top line will shift
subtly when the button and the placeholder trade places. Alas, it seems
infeasible to prevent the gear icon itself from shifting back and forth
without applying more aggressive styling changes.

This is Part 5 of the HiDPI bug fix commit series.