[Frontend] Hide insignificant unread message previews
A preview is *significant* if it has a greater *rank* than any of its
parent previews, where pings of the current user have a rank of 3,
replies to the current user a rank of 2, and all other messages have
a rank of 1. Top-level previews are always shown, their nonexistent
parent having a "rank" of 0.

Also rearranges the code and fixes some bugs.