[Frontend] Avoid undeletable unread messages
Messages posted while the tab is active would never be removed from the
preview pane. This was caused by the preview pane code picking up
incoming messages on its own (instead of being called from the unread
message tracking code as it had been before the undermentioned commit),
filtering the messages on its own, and not doing the latter well. That
strategy was choosen because it promised easy access to the messages'
"notification rank"; however, it turns out that it is even easier to
call the function that computes the rank in the first place instead.

Introduced in commit c50a90e4b4258a9f0985873254dcaf452d17a870, which is
partially reverted.