[Frontend] Fix monospace block markers bug
Now, they require (somewhat) adjacent line breaks again.

Additionally, their type is chosen depending on the actually surrounding
line breaks to reduce mis-detections; this has the questionable side
effect of permitting nested monospace blocks for a first time in
Instant's history.

As quite a few others, this regression was introduced in the ill-fated
commit fe9044a2cf28f5d5aa91de0018a17371de377315.