[Frontend] Revamp window overlay styling
Apparently, it is not possible to exempt scrolling from
pointer-events:none; hence, the entire area of the window overlay
will need to capture mouse events, and thus have a background
similar to that of the popup overlay.

Means of accessing the sidebar (i.e. hide-all and close-all buttons
as for popups, as well as interaction with the various sidebar
display modes) will be added in a later commit.