[Frontend] Remove window-sidebar-drawer interaction
Producing a backdrop for the settings popup that correctly covers
everything below the settings (but not the settings themselves) turns
out to be rather hard when the sidebar is z-indexed. In spite of its
appeal, the re-layering of the sidebar drawer is, therefore, removed.

Partially reverts commit 9b93e52d405cd33e59368ebfdccb1a2eb0beb174.