[Instabot] Avoid race condition on keepalive bot close
When closing the bot from another thread, the keepalive attribute
assignment was not synchronized against the thread running the
main loop; depending on the timing, the main loop thread could reach
line 660 before the assignment was executed, leading to the bot
reconnecting one last time (and potentially sticking around for a
long time).

Now, the close is sequenced after the attribute assignment, which
should eliminate this failure mode.