[Frontend] Fix data message navigation scrolling bugs
The rationale of the jumpTo() change is explained in the pertinent
added comment.

A new and different means of scroll state saving is introduced since
Instant.input.saveScrollState() saves and restores the location of
the input bar, which results in the input bar not moving w.r.t. the
screen when entering/leaving a data message, which is probably not
what the user expects.